[WIP] Change the paddle dealing logic to support proper iambic mode
A proper iambic keyer should allow inserting a 'dit' or a 'dah' to be
played next without needing to wait for the current element to
finish. It also should generate an alternating sequence of 'dits' and
'dahs' when pressing both paddles.

I've changed the paddle reading logic to be interrupt driven by GPIO
change, and added a buffer to set an 'next' element to be played if
another paddle was pressed during the playing of the current element.