Re-write the firmware in C
Got tired of writing PIC assembly by hand, the iambic keyer logic is a
bit more complicated than I expected, with some corner cases that
require more code to make it work.

I'm now using SDCC, it says that PIC using 14 bits and 16 bits
instructions are not fully supported, but it seems to do fine with my
code. It does make gpasm print some annoying warnings though.

Now the iambic logic is complete, it still need some comments to make
more clear how it works.

Still, the keyer speed selection is not implemented yet, it is
hardcoded to '15' for now.